Fix wnd_proc type, memory delete, and i18n check
Adjust SystemTray::wnd_proc return type to long long for correct 64-bit behavior; add missing <cctype> and <cstdint> includes across several files to resolve warnings and ensure proper integer/char types; fix heap free by using delete[] for a Fl_Input* array; replace hardcoded "FAVORITES" string with lang->favorites_btn for localization; remove an unnecessary RadioManager::save_customs() call. These changes address correctness on x64, memory-safety, and i18n/cleanups.

+++ include/SystemTray.h
@@ -17,7 +17,7 @@ class SystemTray {
     void hide_window();
     void show_window();
 
-    static long __stdcall wnd_proc(void*, unsigned int, unsigned long long, long long);
+    static long long __stdcall wnd_proc(void*, unsigned int, unsigned long long, long long);
     static SystemTray* self;
 
 public:
